+++ internal/modules/payment/application/payment_service_rules.go
@@ -142,24 +142,19 @@ func buildOrderSubject(order *orderdomain.Order) string {
 	if order == nil {
 		return ""
 	}
-	if title := firstOrderItemTitle(order.Items); title != "" {
-		return title
+	for i := range order.Items {
+		if title := pickOrderItemTitle(order.Items[i].TitleJSON); title != "" {
+			return title
+		}
 	}
 	for i := range order.Children {
-		if title := firstOrderItemTitle(order.Children[i].Items); title != "" {
-			return title
+		for j := range order.Children[i].Items {
+			if title := pickOrderItemTitle(order.Children[i].Items[j].TitleJSON); title != "" {
+				return title
+			}
 		}
 	}
 	return strings.TrimSpace(order.OrderNo)
-}
-
-func firstOrderItemTitle(items []orderdomain.OrderItem) string {
-	for i := range items {
-		if title := pickOrderItemTitle(items[i].TitleJSON); title != "" {
-			return title
-		}
-	}
-	return ""
 }
 
 func pickOrderItemTitle(title jsonmap.JSON) string {
